+++ BlogposterCMS/mother/modules/pagesManager/config/defaultWidgets.js
@@ -8,21 +8,21 @@ module.exports.DEFAULT_WIDGETS = [
     widgetId: 'pageList',
     widgetType: ADMIN_LANE,
     label: 'Page List',
-    content: '/assets/plainspace/admin/defaultwidgets/pageList.js',
+    content: '/assets/plainspace/widgets/admin/defaultwidgets/pageList.js',
     category: 'core'
   },
   {
     widgetId: 'pageStats',
     widgetType: ADMIN_LANE,
     label: 'Page Stats',
-    content: '/assets/plainspace/admin/defaultwidgets/pageStats.js',
+    content: '/assets/plainspace/widgets/admin/defaultwidgets/pageStats.js',
     category: 'core'
   },
   {
     widgetId: 'pageEditorWidget',
     widgetType: ADMIN_LANE,
     label: 'Page Editor',
-    content: '/assets/plainspace/admin/pageEditorWidgets/pageEditorWidget.js',
+    content: '/assets/plainspace/widgets/admin/pageEditorWidgets/pageEditorWidget.js',
     category: 'core'
   },
 ];